STP-196: Add a custom mocking function that implements support on `document.createElement` in jsdom.
This function is based on a similar solution in geospatial-app. It written in scripts/setup-canvas-jsdom.js and used in the scripts/setup-jsdom.js file.  It simply takes the jsdom version of `document.createElement` and adds a `getContext` property if the `tagLine` given is 'canvas'. JSDOM does not supply this out of the box.