Refactor bar tests for class names.
Using the same component wrapper in multiple tests is problematic, because its state in later tests is affected by the props it was passed in earlier tests. I've refactored these tests to create a fresh wrapper in each. I've also removed redundant assertions, so we only test the behavior described.