Don't use Webpack production mode for bundling demos.
It seems to me the main use of the demos is as an aid to developers. Production mode applies uglification by default, making it far more difficult to use the React devtools because the names of components have been mangled. Uglification also makes builds take longer. Additionally, in production mode, React doesn't print many warnings. These warnings can be very helpful for development. The bundle size is of course quite a bit larger when not using production mode, but I think the tradeoffs are worth it.