Fix failing tests (Scatter)
These tests were failing, I believe, because they depended on Enzyme's ShallowWrapper.setProps completing synchronously. Perhaps that was true in the past, but now at least it seems this method is asynchronous--indeed, it takes an optional callback function to be executed after the update is processed. I've simplified the tests in question so this is no longer an issue.