Revise versioning scripts.
I don't see a real need for the explicit "version" script. We're not actually committing any generated assets from "build:all", since the output directories are ignored by Git. The only reason to run "build:all" prior to bumping the version is to ensure that it completes successfully. To me it makes sense to run "build:all" in the "preversion" step and eliminate the "version" script.