Replace topojson package with topojson-client and topojson-simplify.
As this issue (https://github.com/topojson/topojson/issues/304) on GitHub explains, `npm install` fails with an error if a project tries to include both `topojson` and `topojson-client`.  Because `topojson` is simply a light wrapper for the three subpackages `topojson-client`, `topojson-simplify`, and `topojson-server`, and IHME-UI only uses the first two of these, it seems simplest and most economical to replace the wrapper with the two constituent packages we're actually using.